NFC: Move matmul C matrix promotion related code to Common/
This is in preparation to be reused by the SPIR-V CodeGen pipeline.
