+++ compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUTileAndDistribute.cpp
@@ -7,6 +7,7 @@
 #include "iree-dialects/Dialect/LinalgExt/IR/LinalgExtOps.h"
 #include "iree-dialects/Dialect/LinalgExt/Passes/Transforms.h"
 #include "iree-dialects/Dialect/LinalgExt/Transforms/Transforms.h"
+#include "iree/compiler/Codegen/Common/GPUPatterns.h"
 #include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
 #include "iree/compiler/Codegen/LLVMGPU/KernelConfig.h"
 #include "iree/compiler/Codegen/LLVMGPU/TilingUtils.h"
@@ -27,7 +28,6 @@
 #include "mlir/Support/MathExtras.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 #include "mlir/Transforms/Passes.h"
-#include "mlir/Transforms/SideEffectUtils.h"
 
 using mlir::iree_compiler::IREE::LinalgExt::TilingPatterns;
 
@@ -173,138 +173,6 @@ static void populateTilingToInvocationPatterns(
       context, tilingOptions, f);
 }
 
-static LogicalResult copyToWorkgroupMemory(OpBuilder &b, Value src, Value dst) {
-  Operation *copyOp = b.create<memref::CopyOp>(src.getLoc(), src, dst);
-  setMarker(copyOp, getCopyToWorkgroupMemoryMarker());
-  return success();
-}
-
-using PromotionFilterFunction = std::function<LogicalResult(Operation *op)>;
-
-/// Returns true if op is appropriate contract for promotion.
-static LogicalResult contractOpFilter(Operation *op) {
-  auto linalgOp = dyn_cast<linalg::LinalgOp>(op);
-  if (!linalgOp) return failure();
-  // Limit promotion to matmul and batch matmul, there may be generic
-  // ops with more batch dimensions we didn't distribute and therefore
-  // cannot find a higher bound.
-  return success(linalg::isaContractionOpInterface(op) &&
-                 linalgOp.getNumParallelLoops() >= 2 &&
-                 linalgOp.getNumParallelLoops() <= 3);
-}
-
-template <typename T>
-using LinalgPromotionPattern =
-    mlir::iree_compiler::IREE::LinalgExt::LinalgPromotionPattern<T>;
-static void populatePromotionPatterns(MLIRContext *context,
-                                      RewritePatternSet &patterns,
-                                      PromotionFilterFunction filterFunction,
-                                      ArrayRef<int64_t> operandsToPromote) {
-  patterns.insert<LinalgPromotionPattern<linalg::MatmulOp>,
-                  LinalgPromotionPattern<linalg::BatchMatmulOp>,
-                  LinalgPromotionPattern<linalg::GenericOp>>(
-      context,
-      linalg::LinalgPromotionOptions()
-          .setAllocationDeallocationFns(allocateWorkgroupMemory,
-                                        deallocateWorkgroupMemory)
-          .setCopyInOutFns(copyToWorkgroupMemory, copyToWorkgroupMemory)
-          .setOperandsToPromote(operandsToPromote)
-          .setUseFullTileBuffers({false, false}),
-      IREE::LinalgExt::LinalgTransformationFilter(
-          {StringAttr::get(context, getWorkgroupKTiledMarker())},
-          StringAttr::get(context, getWorkgroupMemoryMarker()))
-          .setMatchByDefault()
-          .addFilter(filterFunction));
-}
-
-static bool propagateCopyDestIntoProducerFill(memref::CopyOp copyOp) {
-  // Look for a fill Op writing into the copyOp source.
-  Operation *prevOp = copyOp->getPrevNode();
-  while (prevOp) {
-    if (isMemoryEffectFree(prevOp)) {
-      prevOp = prevOp->getPrevNode();
-      continue;
-    }
-
-    auto fillOp = dyn_cast<linalg::FillOp>(prevOp);
-    if (!fillOp) break;
-    if (fillOp.output() != copyOp.getSource()) break;
-    // Move the fillOp and change the destination to the copy destination.
-    fillOp->moveBefore(copyOp);
-    fillOp.getOutputsMutable().assign(copyOp.getTarget());
-    return true;
-  }
-  return false;
-}
-
-// Split input/output operand from copy from shared memory into a separate
-// input.
-static void insertInputValueIntoGeneric(Value source, linalg::GenericOp op) {
-  SmallVector<Value> newOperands;
-  SmallVector<AffineMap> maps;
-  for (OpOperand *in : op.getDpsInputOperands()) {
-    newOperands.push_back(in->get());
-    maps.push_back(op.getMatchingIndexingMap(in));
-  }
-  newOperands.push_back(source);
-  assert(op.getNumDpsInits() == 1);
-  OpOperand *outOperand = op.getDpsInitOperand(0);
-  maps.push_back(op.getMatchingIndexingMap(outOperand));
-  maps.push_back(op.getMatchingIndexingMap(outOperand));
-  Location loc = op.getLoc();
-  SmallVector<StringRef> iterTypes(op.getNumLoops(),
-                                   getParallelIteratorTypeName());
-  OpBuilder builder(op);
-  auto newOp = builder.create<linalg::GenericOp>(
-      loc, newOperands, outOperand->get(), maps, iterTypes);
-  newOp.getRegion().getBlocks().splice(newOp.getRegion().begin(),
-                                       op.getRegion().getBlocks());
-
-  Block &payload = newOp.getRegion().front();
-  payload.addArgument(payload.getArguments().back().getType(), loc);
-  setMarker(newOp, getCopyToWorkgroupMemoryMarker());
-}
-
-/// Propagate the shared memory copy into the consumer op if it's a fully
-/// parallel linalg.generic.
-static bool propagateCopySourceIntoConsumerGeneric(
-    memref::CopyOp copyOp, SmallVector<Operation *> &toDelete) {
-  // Look for a generic Op reading the copyOp target.
-  Operation *nextOp = copyOp->getNextNode();
-  while (nextOp) {
-    if (isMemoryEffectFree(nextOp)) {
-      nextOp = nextOp->getNextNode();
-      continue;
-    }
-    auto consumer = dyn_cast<linalg::GenericOp>(nextOp);
-    if (!consumer || consumer.getNumDpsInits() != 1 ||
-        !consumer.getMatchingIndexingMap(consumer.getDpsInitOperand(0))
-             .isIdentity())
-      break;
-    if (*consumer.getOutputs().begin() != copyOp.getTarget()) break;
-    insertInputValueIntoGeneric(copyOp.getSource(), consumer);
-    toDelete.push_back(consumer);
-    return true;
-  }
-  return false;
-}
-
-/// Transformation to propagate FillOp + CopyOp to temp allocation.
-/// This is needed because we are doing promotion to shared memory on buffers.
-/// This is a fragile and temporary solution until we move to be able to do this
-/// kind of transformations on tensors.
-static void propagateSharedMemCopy(func::FuncOp funcOp) {
-  SmallVector<Operation *> toDelete;
-  funcOp.walk([&toDelete](memref::CopyOp copyOp) {
-    if (hasMarker(copyOp, getCopyToWorkgroupMemoryMarker())) {
-      if (propagateCopyDestIntoProducerFill(copyOp) ||
-          propagateCopySourceIntoConsumerGeneric(copyOp, toDelete))
-        toDelete.push_back(copyOp.getOperation());
-    }
-  });
-  for (Operation *op : toDelete) op->erase();
-}
-
 namespace {
 struct LLVMGPUTileAndDistributePass
     : public LLVMGPUTileAndDistributeBase<LLVMGPUTileAndDistributePass> {
@@ -327,13 +195,12 @@ struct LLVMGPUTileAndDistributePass
     // allocation. This needs to be done before reduction tiling.
     {
       RewritePatternSet promotionPatterns(&getContext());
-      populatePromotionPatterns(context, promotionPatterns, contractOpFilter,
-                                {2});
+      populateContractPromotionPatterns(promotionPatterns, {2});
       if (failed(applyPatternsAndFoldGreedily(funcOp,
                                               std::move(promotionPatterns)))) {
         return signalPassFailure();
       }
-      propagateSharedMemCopy(funcOp);
+      propagateSharedMemoryCopy(funcOp);
     }
 
     // Tile again at the workgroup level since reduction dimension were
@@ -358,30 +225,14 @@ struct LLVMGPUTileAndDistributePass
     if (flatWorkgroupSize > kWarpSize) {
       RewritePatternSet promotionPatterns(&getContext());
 
-      populatePromotionPatterns(context, promotionPatterns, contractOpFilter,
-                                {0, 1});
+      populateContractPromotionPatterns(promotionPatterns, {0, 1});
 
       if (failed(applyPatternsAndFoldGreedily(funcOp,
                                               std::move(promotionPatterns)))) {
         return signalPassFailure();
       }
-      // Insert barriers before and after copies to workgroup memory and skip
-      // insert barriers between back to back copy to workgroup memory.
-      OpBuilder builder(&getContext());
-      funcOp.walk([&builder](Operation *copyOp) {
-        if (hasMarker(copyOp, getCopyToWorkgroupMemoryMarker())) {
-          Operation *prevOp = copyOp->getPrevNode();
-          if (!prevOp || !hasMarker(prevOp, getCopyToWorkgroupMemoryMarker())) {
-            builder.setInsertionPoint(copyOp);
-            builder.create<gpu::BarrierOp>(copyOp->getLoc());
-          }
-          Operation *nextOp = copyOp->getNextNode();
-          if (!nextOp || !hasMarker(nextOp, getCopyToWorkgroupMemoryMarker())) {
-            builder.setInsertionPointAfter(copyOp);
-            builder.create<gpu::BarrierOp>(copyOp->getLoc());
-          }
-        }
-      });
+      // Insert barriers before and after copies to workgroup memory.
+      insertBarriersAroundSharedMemoryCopy(funcOp);
     }
 
     {
