+++ compiler/src/iree/compiler/Codegen/Utils/GPUUtils.h
@@ -11,46 +11,66 @@
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
 
 namespace mlir {
-namespace memref {
-class SubViewOp;
-}  // namespace memref
-
 namespace iree_compiler {
 
 static constexpr int32_t kNumGPUDims = 3;
 static constexpr int32_t kWarpSize = 32;
 
+//===----------------------------------------------------------------------===//
+// GPU processor IDs and sizes
+//===----------------------------------------------------------------------===//
+
 llvm::SmallVector<linalg::ProcInfo, 2> getGPUThreadIdsAndCounts(
     OpBuilder &builder, Location loc, unsigned numDims,
     llvm::ArrayRef<int64_t> workgroupSize);
 
-/// Compute subgroup ID. CUDA doesn't have a subgroupId equivalent so we are are
-/// computing the subgroup ID based on the threadID.
-/// When tiling to warp we assume each warp is full and we pick a workgroup
-/// size so that `workgroupSize.x % warpSize == 0`. This is why we can have
-/// warpId = { threadId.x / warpSize, threadId.y, threadId.z }.
+/// Computes subgroup ID and returns in (X, Y, Z) order.
+///
+/// Note that CUDA doesn't have a subgroupId equivalent so we are are computing
+/// the subgroup ID based on the threadID. When tiling to warp we assume each
+/// warp is full and we pick a workgroup size so that `workgroupSize.x %
+/// warpSize == 0`. This is why we can have warpId = { threadId.x / warpSize,
+/// threadId.y, threadId.z }.
 llvm::SmallVector<linalg::ProcInfo, 2> getSubgroupIdsAndCounts(
     OpBuilder &builder, Location loc, unsigned warpSize, unsigned numDims,
     llvm::ArrayRef<int64_t> numSubgroups);
 
-/// Return the workgroup size associated to the funcOp entry point.
+/// Returns the workgroup size associated to the funcOp entry point.
 std::array<int64_t, 3> getWorkgroupSize(func::FuncOp funcOp);
 
-/// Return true if we can use all threads to perform vectorized load/store of
+//===----------------------------------------------------------------------===//
+// GPU vectorization
+//===----------------------------------------------------------------------===//
+
+/// Returns true if we can use all threads to perform vectorized load/store of
 /// the given `shape`.
 bool canPerformVectorAccessUsingAllThreads(ArrayRef<int64_t> shape,
                                            int64_t threadCount,
                                            int64_t vectorSize);
 
-/// Allocate GPU workgroup memory matching the given `subview`. If there are
+//===----------------------------------------------------------------------===//
+// GPU workgroup memory
+//===----------------------------------------------------------------------===//
+
+/// Allocates GPU workgroup memory matching the given `subview`. If there are
 /// dynamic dimensions, the bounds are in `sizeBounds`.
 Optional<Value> allocateWorkgroupMemory(OpBuilder &builder,
                                         memref::SubViewOp subview,
                                         ArrayRef<Value> sizeBounds,
                                         DataLayout &);
 
-/// Deallocate GPU workgroup memory behind `buffer`.
+/// Deallocates GPU workgroup memory behind `buffer`.
 LogicalResult deallocateWorkgroupMemory(OpBuilder &, Value buffer);
+
+/// Copies `src` value to `dst` in shared memory.
+LogicalResult copyToWorkgroupMemory(OpBuilder &builder, Value src, Value dst);
+
+/// Propagates shared memory copy to producer linalg.fill or consumer
+/// linalg.generic when possible.
+void propagateSharedMemoryCopy(func::FuncOp funcOp);
+
+/// Inserts barriers before and after shared memory copy.
+void insertBarriersAroundSharedMemoryCopy(func::FuncOp funcOp);
 
 }  // namespace iree_compiler
 }  // namespace mlir
